requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-filter==25.1
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
email_validator==2.2.0
idna==3.10
mysqlclient==2.2.7
phonenumbers==9.0.0
pillow==11.1.0
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
